New Member
Function will take in an array of multiple arrays
where the first number in the array indicates age
and the second number indicates their handicap score.
In order to be considered for a senior membership,
a member needs to be at least 55 years old and have
a handicap greater than 7. Otherwise, they're placed
in the Open membership plan.