Title Case
Create a function that takes in two arguments, a title
and minorWords, where your task is to capitalize the first
letter of a word as if it were a movie or book title.
The minor words play into this by delcaring that if the
title has words that are in the minor words list, do not
capitalize them. The first letter of the first word in
any title must be capitalized. If minor words aren't declared,
capitalize first letter of all words. If title is an empty string
will return a empty string back.